Move GetBalances into utils package
Move GetBalances into utils package as it's not a contract, it's utilities to work on contract states.
