+++ samples/trader-demo/src/main/kotlin/net/corda/traderdemo/TraderDemoClientApi.kt
@@ -13,7 +13,7 @@ import net.corda.finance.DOLLARS
 import net.corda.finance.USD
 import net.corda.finance.contracts.asset.CommercialPaper
 import net.corda.finance.contracts.asset.Cash
-import net.corda.finance.contracts.getCashBalance
+import net.corda.finance.utils.getCashBalance
 import net.corda.finance.flows.CashIssueFlow
 import net.corda.finance.flows.CashPaymentFlow
 import net.corda.node.services.vault.VaultSchemaV1
