+++ samples/trader-demo/src/main/kotlin/net/corda/traderdemo/flow/BuyerFlow.kt
@@ -12,7 +12,7 @@ import net.corda.core.transactions.SignedTransaction
 import net.corda.core.utilities.ProgressTracker
 import net.corda.core.utilities.unwrap
 import net.corda.finance.contracts.asset.CommercialPaper
-import net.corda.finance.contracts.getCashBalances
+import net.corda.finance.utils.getCashBalances
 import net.corda.finance.flows.TwoPartyTradeFlow
 import java.util.*
